+++ app/src/main/java/no/ntnu/klubbhuset/ui/login/LoginViewModelFactory.java
@@ -6,9 +6,6 @@ import androidx.lifecycle.ViewModel;
 import androidx.lifecycle.ViewModelProvider;
 import androidx.annotation.NonNull;
 
-import no.ntnu.klubbhuset.data.LoginDataSource;
-import no.ntnu.klubbhuset.data.LoginRepository;
-
 /**
  * ViewModel provider factory to instantiate LoginViewModel.
  * Required given LoginViewModel has a non-empty constructor
